+++ agent/bindings/kotlin/lib/src/main/kotlin/com/sourcegraph/cody/protocol_generated/ChatExportResult.kt
@@ -2,7 +2,7 @@
 package com.sourcegraph.cody.protocol_generated
 
 data class ChatExportResult(
-  val chatID: String? = null,
-  val transcript: SerializedChatTranscript? = null,
+  val chatID: String,
+  val transcript: SerializedChatTranscript,
 )