+++ agent/bindings/kotlin/lib/src/main/kotlin/com/sourcegraph/cody/protocol_generated/Chat_ModelsParams.kt
@@ -2,6 +2,6 @@
 package com.sourcegraph.cody.protocol_generated
 
 data class Chat_ModelsParams(
-  val modelUsage: ModelUsage? = null, // Oneof: chat, edit
+  val modelUsage: ModelUsage, // Oneof: chat, edit
 )