+++ agent/bindings/kotlin/lib/src/main/kotlin/com/sourcegraph/cody/protocol_generated/CodyTaskState.kt
@@ -2,14 +2,14 @@
 package com.sourcegraph.cody.protocol_generated
 
 data class CodyTaskState(
-  val idle: Int? = null,
-  val working: Int? = null,
-  val inserting: Int? = null,
-  val applying: Int? = null,
-  val formatting: Int? = null,
-  val applied: Int? = null,
-  val finished: Int? = null,
-  val error: Int? = null,
-  val pending: Int? = null,
+  val idle: Int,
+  val working: Int,
+  val inserting: Int,
+  val applying: Int,
+  val formatting: Int,
+  val applied: Int,
+  val finished: Int,
+  val error: Int,
+  val pending: Int,
 )