+++ agent/bindings/kotlin/lib/src/main/kotlin/com/sourcegraph/cody/protocol_generated/DebugMessage.kt
@@ -2,7 +2,7 @@
 package com.sourcegraph.cody.protocol_generated
 
 data class DebugMessage(
-  val channel: String? = null,
-  val message: String? = null,
+  val channel: String,
+  val message: String,
 )