+++ agent/bindings/kotlin/lib/src/main/kotlin/com/sourcegraph/cody/protocol_generated/Graphql_GetRepoIdsParams.kt
@@ -2,7 +2,7 @@
 package com.sourcegraph.cody.protocol_generated
 
 data class Graphql_GetRepoIdsParams(
-  val names: List<String>? = null,
-  val first: Int? = null,
+  val names: List<String>,
+  val first: Int,
 )