+++ agent/bindings/kotlin/lib/src/main/kotlin/com/sourcegraph/cody/protocol_generated/OptionsParams.kt
@@ -2,7 +2,7 @@
 package com.sourcegraph.cody.protocol_generated
 
 data class OptionsParams(
-  val undoStopBefore: Boolean? = null,
-  val undoStopAfter: Boolean? = null,
+  val undoStopBefore: Boolean,
+  val undoStopAfter: Boolean,
 )