+++ agent/bindings/kotlin/lib/src/main/kotlin/com/sourcegraph/cody/protocol_generated/ProgressReportParams.kt
@@ -2,7 +2,7 @@
 package com.sourcegraph.cody.protocol_generated
 
 data class ProgressReportParams(
-  val id: String? = null,
+  val id: String,
   val message: String? = null,
   val increment: Int? = null,
 )