+++ agent/bindings/kotlin/lib/src/main/kotlin/com/sourcegraph/cody/protocol_generated/ProgressStartParams.kt
@@ -2,7 +2,7 @@
 package com.sourcegraph.cody.protocol_generated
 
 data class ProgressStartParams(
-  val id: String? = null,
-  val options: ProgressOptions? = null,
+  val id: String,
+  val options: ProgressOptions,
 )