+++ agent/bindings/kotlin/lib/src/main/kotlin/com/sourcegraph/cody/protocol_generated/RemoteRepo_HasParams.kt
@@ -2,6 +2,6 @@
 package com.sourcegraph.cody.protocol_generated
 
 data class RemoteRepo_HasParams(
-  val repoName: String? = null,
+  val repoName: String,
 )