+++ agent/src/index.test.ts
@@ -11,7 +11,7 @@ import { ModelUsage, isWindows } from '@sourcegraph/cody-shared'
 
 import { URI } from 'vscode-uri'
 import type { RequestMethodName } from '../../vscode/src/jsonrpc/jsonrpc'
-import { TestClient, asTranscriptMessage } from './TestClient'
+import { TestClient, asTranscriptMessage, getAgentDir } from './TestClient'
 import { decodeURIs } from './decodeURIs'
 import { isNode16 } from './isNode16'
 import type {
@@ -64,7 +64,18 @@ describe('Agent', () => {
         return
     }
 
-    const client = new TestClient({
+    // Bundle the agent. When running `pnpm run test`, vitest doesn't re-run this step.
+    //
+    // ⚠️ If this line fails when running unit tests, chances are that the error is being swallowed.
+    // To see the full error, run this file in isolation:
+    //
+    //   pnpm test agent/src/index.test.ts
+    execSync('pnpm run build:agent', {
+        cwd: getAgentDir(),
+        stdio: 'inherit',
+    })
+
+    const client = TestClient.create({
         name: 'defaultClient',
         // The redacted ID below is copy-pasted from the recording file and
         // needs to be updated whenever we change the underlying access token.
@@ -73,17 +84,6 @@ describe('Agent', () => {
         accessToken:
             process.env.SRC_ACCESS_TOKEN ??
             'REDACTED_b09f01644a4261b32aa2ee4aea4f279ba69a57cff389f9b119b5265e913c0ea4',
-    })
-
-    // Bundle the agent. When running `pnpm run test`, vitest doesn't re-run this step.
-    //
-    // ⚠️ If this line fails when running unit tests, chances are that the error is being swallowed.
-    // To see the full error, run this file in isolation:
-    //
-    //   pnpm test agent/src/index.test.ts
-    execSync('pnpm run build:agent', {
-        cwd: client.getAgentDir(),
-        stdio: 'inherit',
     })
 
     // Initialize inside beforeAll so that subsequent tests are skipped if initialization fails.
@@ -112,10 +112,10 @@ describe('Agent', () => {
         expect(valid?.isLoggedIn).toBeTruthy()
 
         // Confirm .cody/ignore is active at start up
-        const codyIgnore = await client.request('check/isCodyIgnoredFile', {
-            urls: [ignoredPath],
+        const ignore = await client.request('ignore/test', {
+            uri: URI.file(ignoredPath).toString(),
         })
-        expect(codyIgnore).toBeTruthy()
+        expect(ignore.policy).toBe('ignore')
     }, 10_000)
 
     beforeEach(async () => {
@@ -542,10 +542,10 @@ describe('Agent', () => {
             const codyIgnoreConfigFile = client.workspace.getDocument(codyIgnoreConfig)
             expect(codyIgnoreConfigFile?.content).toBeDefined()
 
-            const result = await client.request('check/isCodyIgnoredFile', {
-                urls: [ignoredPath],
+            const result = await client.request('ignore/test', {
+                uri: URI.file(ignoredPath).toString(),
             })
-            expect(result).toBeTruthy()
+            expect(result.policy).toBe('ignore')
         }, 10_000)
 
         it('autocomplete/execute on ignored file', async () => {
@@ -578,11 +578,13 @@ describe('Agent', () => {
             // Current file which is ignored, should not be included in context files
             expect(contextFiles.find(f => f.uri.toString() === ignoredUri.toString())).toBeUndefined()
             // Ignored file should not be included in context files
-            const contextFilesUrls = contextFiles.map(f => f.uri?.path)
-            const result = await client.request('check/isCodyIgnoredFile', {
-                urls: contextFilesUrls,
-            })
-            expect(result).toBeFalsy()
+            const contextFilesUrls = contextFiles.map(f => f.uri).filter(uri => uri)
+            const result = await Promise.all(
+                contextFilesUrls.map(uri => client.request('ignore/test', { uri: uri.toString() }))
+            )
+            for (const r of result) {
+                expect(r.policy).toBe('use')
+            }
             // Files that are not ignored should be used as context files
             expect(contextFiles.length).toBeGreaterThan(0)
         }, 30_000)
@@ -604,10 +606,14 @@ describe('Agent', () => {
             // Since no enhanced context is requested, no context files should be included
             expect(contextFiles.length).toBe(0)
             // Ignored file should not be included in context files
-            const result = await client.request('check/isCodyIgnoredFile', {
-                urls: contextUrls,
-            })
-            expect(result).toBeFalsy()
+            const result = await Promise.all(
+                contextUrls.map(uri =>
+                    client.request('ignore/test', {
+                        uri,
+                    })
+                )
+            )
+            expect(result.every(entry => entry.policy === 'use')).toBe(true)
         }, 30_000)
 
         it('chat command on an ignored file', async () => {
@@ -633,19 +639,19 @@ describe('Agent', () => {
 
         it('ignore rule is not case sensitive', async () => {
             const alsoIgnoredPath = path.join(workspaceRootPath, 'src/is_ignored.ts')
-            const result = await client.request('check/isCodyIgnoredFile', {
-                urls: [alsoIgnoredPath],
+            const result = await client.request('ignore/test', {
+                uri: URI.file(alsoIgnoredPath).toString(),
             })
-            expect(result).toBeTruthy()
+            expect(result.policy).toBe('ignore')
         })
 
         afterAll(async () => {
             // Makes sure cody ignore is still active after tests
             // as it should stay active for each workspace session.
-            const result = await client.request('check/isCodyIgnoredFile', {
-                urls: [ignoredPath],
+            const result = await client.request('ignore/test', {
+                uri: URI.file(ignoredPath).toString(),
             })
-            expect(result).toBeTruthy()
+            expect(result.policy).toBe('ignore')
 
             // Check the network requests to ensure no requests include context from ignored files
             const { requests } = await client.request('testing/networkRequests', null)
@@ -1363,7 +1369,7 @@ describe('Agent', () => {
     })
 
     describe('RateLimitedAgent', () => {
-        const rateLimitedClient = new TestClient({
+        const rateLimitedClient = TestClient.create({
             name: 'rateLimitedClient',
             accessToken:
                 process.env.SRC_ACCESS_TOKEN_WITH_RATE_LIMIT ??
@@ -1392,7 +1398,7 @@ describe('Agent', () => {
     })
 
     describe('Enterprise', () => {
-        const enterpriseClient = new TestClient({
+        const enterpriseClient = TestClient.create({
             name: 'enterpriseClient',
             accessToken:
                 process.env.SRC_ENTERPRISE_ACCESS_TOKEN ??
@@ -1545,6 +1551,43 @@ describe('Agent', () => {
             expect(codyForDos.result).toBe(false)
         })
 
+        it('testing/ignore/overridePolicy', async () => {
+            // To test that testing/ignore/overridePolicy generates an
+            // ignore/didChange notification, 'stop' sets up a Promise, 'go'
+            // unblocks it.
+            let go = () => {}
+            function stop() {
+                return new Promise(resolve => {
+                    go = () => resolve(undefined)
+                })
+            }
+            enterpriseClient.registerNotification('ignore/didChange', () => {
+                go()
+            })
+            expect(
+                await enterpriseClient.request('ignore/test', { uri: 'file:///foo/bar.txt' })
+            ).toStrictEqual({ policy: 'use' })
+            await Promise.all([
+                stop(),
+                enterpriseClient.request('testing/ignore/overridePolicy', {
+                    repoRe: '$^',
+                    uriRe: '.*bar.*',
+                }),
+            ])
+            expect(
+                await enterpriseClient.request('ignore/test', { uri: 'file:///foo/bar.txt' })
+            ).toStrictEqual({ policy: 'ignore' })
+            expect(
+                await enterpriseClient.request('ignore/test', { uri: 'file:///foo/quux.txt' })
+            ).toStrictEqual({ policy: 'use' })
+            await Promise.all([stop(), enterpriseClient.request('testing/ignore/overridePolicy', null)])
+            expect(
+                await enterpriseClient.request('ignore/test', { uri: 'file:///foo/bar.txt' })
+            ).toStrictEqual({ policy: 'use' })
+            // Stop listening to 'ignore/didChange'
+            enterpriseClient.registerNotification('ignore/didChange', () => {})
+        })
+
         afterAll(async () => {
             const { requests } = await enterpriseClient.request('testing/networkRequests', null)
             const nonServerInstanceRequests = requests
@@ -1559,7 +1602,7 @@ describe('Agent', () => {
     // Enterprise tests are run at demo instance, which is at a recent release version.
     // Use this section if you need to run against S2 which is released continuously.
     describe('Enterprise - close main branch', () => {
-        const enterpriseClient = new TestClient({
+        const enterpriseClient = TestClient.create({
             name: 'enterpriseMainBranchClient',
             accessToken:
                 process.env.SRC_S2_ACCESS_TOKEN ??