+++ vscode/src/jsonrpc/agent-protocol.ts
@@ -123,10 +123,6 @@ export type ClientRequests = {
 
     'graphql/getRepoIdIfEmbeddingExists': [{ repoName: string }, string | null]
     'graphql/getRepoId': [{ repoName: string }, string | null]
-    /**
-     * Checks if a given set of URLs includes a Cody ignored file.
-     */
-    'check/isCodyIgnoredFile': [{ urls: string[] }, boolean]
 
     'git/codebaseName': [{ url: string }, string | null]
 
@@ -179,6 +175,20 @@ export type ClientRequests = {
             limitHit: boolean
         },
     ]
+
+    // Gets whether the specified URI is sensitive and should not be sent to
+    // LLM providers.
+    'ignore/test': [
+        { uri: string },
+        {
+            policy: 'ignore' | 'use'
+        },
+    ]
+
+    // For testing. Overrides any ignore policy to ignore repositories and URIs
+    // which match the specified regular expressions. Pass `undefined` to remove
+    // the override.
+    'testing/ignore/overridePolicy': [{ repoRe: string; uriRe: string } | null, null]
 
     // Gets whether the specific repo name is known on the remote.
     'remoteRepo/has': [{ repoName: string }, { result: boolean }]
@@ -313,6 +323,10 @@ export type ServerNotifications = {
 
     'codeLenses/display': [DisplayCodeLensParams]
 
+    // The set of ignored files/repositories has changed. The client should
+    // re-query using ignore/test.
+    'ignore/didChange': [null]
+
     // Low-level webview notification for the given chat session ID (created via
     // chat/new). Subscribe to these messages to get access to streaming updates
     // on the chat reply.
@@ -329,8 +343,7 @@ export type ServerNotifications = {
 
     // The list of remote repositories changed. Results from remoteRepo/list
     // may be stale and should be requeried.
-    // biome-ignore lint/complexity/noBannedTypes: May add details about the change later.
-    'remoteRepo/didChange': [{}]
+    'remoteRepo/didChange': [null]
     // Reflects the state of fetching the repository list. After fetching is
     // complete, or errored, the results from remoteRepo/list will not change.
     // When configuration changes, repo fetching may re-start.